fix: only attempt to delete from S3 if an url is present

+++ pkg/database/service.go
@@ -126,15 +126,17 @@ func (s service) Delete(id uint) error {
 		return err
 	}
 
-	u, err := url.Parse(d.Url)
-	if err != nil {
-		return err
-	}
+	if d.Url != "" {
+		u, err := url.Parse(d.Url)
+		if err != nil {
+			return err
+		}
 
-	key := u.Path[1:] // Strip leading "/"
-	err = s.s3Client.Delete(s.c.Bucket, key)
-	if err != nil {
-		return err
+		key := u.Path[1:] // Strip leading "/"
+		err = s.s3Client.Delete(s.c.Bucket, key)
+		if err != nil {
+			return err
+		}
 	}
 
 	return s.repository.Delete(id)
